+++ app/integration_test/tests/bug_reporter.dart
@@ -1,7 +1,7 @@
 import 'package:acter/features/bug_report/actions/open_bug_report.dart';
 import 'package:acter/features/bug_report/pages/bug_report_page.dart';
 import 'package:acter/features/home/data/keys.dart';
-import 'package:acter/config/app_shell.dart';
+import 'package:acter/features/main/app_shell.dart';
 import 'package:acter/features/search/model/keys.dart';
 import 'package:acter/router/router.dart';
 import 'package:convenient_test_dev/convenient_test_dev.dart';
