+++ app/lib/router/router.dart
@@ -2,7 +2,7 @@ import 'package:acter/common/extensions/options.dart';
 import 'package:acter/common/pages/not_found.dart';
 import 'package:acter/common/utils/constants.dart';
 import 'package:acter/common/utils/routes.dart';
-import 'package:acter/config/app_shell.dart';
+import 'package:acter/features/main/app_shell.dart';
 import 'package:acter/features/home/providers/client_providers.dart';
 import 'package:acter/router/general_router.dart';
 import 'package:acter/router/shell_routers/activities_shell_router.dart';
