+++ app/lib/router/utils.dart
@@ -1,6 +1,6 @@
 import 'package:acter/common/providers/chat_providers.dart';
 import 'package:acter/common/utils/routes.dart';
-import 'package:acter/config/app_shell.dart';
+import 'package:acter/features/main/app_shell.dart';
 import 'package:acter/config/setup.dart';
 import 'package:acter/router/providers/router_providers.dart';
 import 'package:acter/router/router.dart';
